Refactor UserMutations validation logic into static classes
This commit involves a significant refactor of the user-related operations within UserMutations.cs. Two new static classes – AppUserInsertDtoHandling and AppUserLoginDtoHandling – have been introduced to handle validation logic that was previously embedded within the CreateUser and SignIn methods. This has served to streamline the main methods and centralize the validation responsibility.