Add user deletion function and implement CORS
A function to delete users has been implemented in UserRepository, along with CORS in API project. The deletion function removes a specific user based on the provided id from the database. A test for the functionality has also been adjusted to accommodate this change. The implementation of CORS (Cross-Origin Resource Sharing) resolves potential browser security issues in accessing the API from different domain.