Remove justification handling with files and implement FTP
The files handling for justifications has been removed from the project. The previous implementation was based on file storage directly in the database. It has been replaced with an FTP implementation that stores justification files on an FTP server. Also, files are now received and served through HTTP with expirable tokens. Dependency on Files.EntityFrameworkCore.Extensions has been removed.