Add feature to change justification file extension
Introduced a new feature to handle the change of justification file extensions. This includes adding a property to handle and store justification extensions in the Expense model, as well as a method to change the justification extension in the Expense repository. In addition, this also extracts the extension logic from the GroupMutations class to the newly created ExpenseMutations class. The JustificationsController was also updated to incorporate new changes.